Refactor - move non-wordpress specific views in own component
